+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 
 
 ### Improvements
+- Add `CoCalculator` Class for co-occurrence topic & keyword
+- Add 5 Step to export dataset
 
 ### Bug Fixes